requirements.txt
Django==1.11
celery==4.0.2
dj-database-url==0.4.1
django-bootstrap-form==3.2.1
django-autocomplete-light==3.1.8
django-crontab==0.7.1
django-crispy-forms==1.6.1
django-floppyforms==1.7.0
django-formtools==1.0
django-ordered-model==1.3.0
django-suit==0.2.20
djangorestframework==3.5.4
django-daterange-filter==1.1.1
django-filter==0.10.0
django-import-export==0.5.1
django-nested-admin==3.0.16
django-push-notifications==1.5.0
django-registration-redux==1.4
easy-thumbnails==2.3
gevent==1.1.1
greenlet==0.4.10
gunicorn==19.6.0
Markdown==2.6.2
openpyxl==2.4.2
Pillow==3.3.0
requests==2.10.0
sendgrid==2.2.1
sendgrid-django==2.0.0
six==1.10.0
tablib==0.10.0
wheel==0.24.0
xlwt==1.0.0
XlsxWriter==0.9.3